refactor: Rename react module to rnSetup module
