+++ app/build.gradle
@@ -60,7 +60,7 @@ dependencies {
     implementation "com.facebook.react:react-native:0.66.1"
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.9")
     implementation "androidx.fragment:fragment-ktx:1.3.6"
-    implementation project(":react")
+    implementation project(':rnSetup')
 
     if (enableHermes) {
         def hermesPath = "../reactNative/node_modules/hermes-engine/android/"
