+++ rnSetup/.gitignore
