+++ rnSetup/src/main/java/com/betatech/rnSetup/MyReactFragment.kt
@@ -1,4 +1,4 @@
-package com.betatech.react
+package com.betatech.rnSetup
 
 import android.app.Application
 import android.os.Bundle
